test: move setup to cli
