+++ test/cli/fixtures/setup-files/console-setup.ts
