--- test/setup/tests/empty-setup.test.ts
@@ -1,5 +1,0 @@
-import { expect, it } from 'vitest'
-
-it('should success', async () => {
-  expect(1 + 1).toBe(2)
-})
